requirements.txt
pandas
beautifulsoup4
requests
python-telegram-bot
asyncio
matplotlib
lxml